requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
pandas==2.2.1
numpy==1.26.4
python-multipart==0.0.9
streamlit==1.33.0
google-genai
